much progrsss on connfiguration

+++ main-module/src/main/java/org/opennms/plugins/messagenotifier/Controller.java
@@ -2,6 +2,7 @@ package org.opennms.plugins.messagenotifier;
 
 import java.io.File;
 import java.util.ArrayList;
+import java.util.Arrays;
 import java.util.HashMap;
 import java.util.LinkedHashMap;
 import java.util.List;
@@ -21,6 +22,7 @@ import org.opennms.plugins.mqtt.config.MessageDataParserConfig;
 import org.opennms.plugins.mqtt.config.MQTTClientConfig;
 import org.opennms.plugins.mqtt.config.MQTTReceiverConfig;
 import org.opennms.plugins.mqtt.config.MessageClientConfig;
+import org.opennms.plugins.mqtt.config.MessageParserConfig;
 import org.opennms.plugins.mqttclient.MQTTClientImpl;
 import org.opennms.plugins.mqttclient.NodeByForeignSourceCacheImpl;
 import org.opennms.protocols.xml.config.XmlGroups;
@@ -30,24 +32,47 @@ import org.slf4j.LoggerFactory;
 public class Controller {
 	private static final Logger LOG = LoggerFactory.getLogger(Controller.class);
 
-	private NodeByForeignSourceCacheImpl m_nodeByForeignSourceCacheImpl=null;
-	
-	private MqttEventNotificationClient m_mqttEventNotificationClient=null;
+	// internal map of client name / mqtt client
+	private Map<String,MQTTClientImpl> m_clientMap= new HashMap<String, MQTTClientImpl>();
 
+	// internal map of client name / mqttrx service
+	private Map<String,MqttRxService>  m_rxServiceMap= new HashMap<String, MqttRxService>();
+	
+	// internal parsed xml configuration
+	private MQTTReceiverConfig m_MQTTReceiverConfig = null;
+	
+	/*
+	 * externally set class properties
+	 */
+	
+	// url to load config
+	private String m_configFile=null;
+	
+	// notification message handler - convert messages to events or data
+	private NotificationMessageHandler m_notificationMessageHandler=null;
+	
+	// node cache for opennms nodes
+    private NodeByForeignSourceCacheImpl m_nodeByForeignSourceCacheImpl=null;
+	
 	// m_messageReceiverServices are used to define message notifiers 
 	// within the initial blueprint. Used for ReST interface
 	private List<MqttRxService> m_messageReceiverServices = null;
 
-	private MQTTReceiverConfig m_MQTTReceiverConfig = null;
-
-	private Map<String,MQTTClientImpl> m_clientMap= new HashMap<String, MQTTClientImpl>();
-
-	private Map<String,MqttRxService>  m_rxServiceMap= new HashMap<String, MqttRxService>();
-
-	private String m_configFile=null;
-
+	// message queue and consumer thread pool
 	private MessageNotificationClientQueueImpl m_messageNotificationClientQueueImpl = null;
 
+
+	/*
+	 * getters and setters for class properties
+	 */
+	
+	public NotificationMessageHandler getNotificationMessageHandler() {
+		return m_notificationMessageHandler;
+	}
+
+	public void setNotificationMessageHandler(NotificationMessageHandler notificationMessageHandler) {
+		this.m_notificationMessageHandler = notificationMessageHandler;
+	}
 
 	public NodeByForeignSourceCacheImpl getNodeByForeignSourceCacheImpl() {
 		return m_nodeByForeignSourceCacheImpl;
@@ -56,14 +81,6 @@ public class Controller {
 	public void setNodeByForeignSourceCacheImpl(
 			NodeByForeignSourceCacheImpl nodeByForeignSourceCacheImpl) {
 		this.m_nodeByForeignSourceCacheImpl = nodeByForeignSourceCacheImpl;
-	}
-
-	public MqttEventNotificationClient getMqttEventNotificationClient() {
-		return m_mqttEventNotificationClient;
-	}
-
-	public void setMqttEventNotificationClient(	MqttEventNotificationClient mqttEventNotificationClient) {
-		this.m_mqttEventNotificationClient = mqttEventNotificationClient;
 	}
 
 	public List<MqttRxService> getMessageReceiverServices() {
@@ -82,7 +99,20 @@ public class Controller {
 		this.m_configFile = configFile;
 	}
 
-	public void init(){
+	public MessageNotificationClientQueueImpl getMessageNotificationClientQueueImpl() {
+		return m_messageNotificationClientQueueImpl;
+	}
+
+	public void setMessageNotificationClientQueueImpl(
+			MessageNotificationClientQueueImpl messageNotificationClientQueueImpl) {
+		this.m_messageNotificationClientQueueImpl = messageNotificationClientQueueImpl;
+	}
+
+	/*
+	 * business methods
+	 */
+	
+	public void loadConfig(){
 
 		m_MQTTReceiverConfig = loadConfigFile();
 
@@ -93,24 +123,27 @@ public class Controller {
 		if(m_MQTTReceiverConfig.getCreateDummyInterfaces()!=null) m_nodeByForeignSourceCacheImpl.setCreateDummyInterfaces(m_MQTTReceiverConfig.getCreateDummyInterfaces());
 		if(m_MQTTReceiverConfig.getCreateMissingNodes()!=null) m_nodeByForeignSourceCacheImpl.setCreateMissingNodes(m_MQTTReceiverConfig.getCreateMissingNodes());
 		if(m_MQTTReceiverConfig.getCreateNodeAssetData()!=null) m_nodeByForeignSourceCacheImpl.setcreateNodeAssetData(m_MQTTReceiverConfig.getCreateNodeAssetData());
-		//m_nodeByForeignSourceCacheImpl.init(); // TODO
-		//m_nodeByForeignSourceCacheImpl.destroy(); // TODO
+
 
 		// Set Up mqtt clients and message receivers
+		
 		// set up mqtt receivers
 		for(MQTTClientConfig mqttConfig:m_MQTTReceiverConfig.getMqttClients()){
 			if (m_clientMap.containsKey(mqttConfig.getClientInstanceId())) 
-				throw new RuntimeException("duplicate ClientInstanceId '"+mqttConfig.getClientInstanceId()
+				throw new RuntimeException("duplicate mqtt receiver ClientInstanceId '"+mqttConfig.getClientInstanceId()
 						+ "' in configuration");
 			MQTTClientImpl client = new MQTTClientImpl(mqttConfig);
 			m_clientMap.put(mqttConfig.getClientInstanceId(), client);
 		}
 
-		// set up other receivers
+		// set up message receivers
 		Map<String, MqttRxService> messageReceivers = new HashMap<String,MqttRxService>();
 		if(m_messageReceiverServices!=null){
 			for(MqttRxService messageReceiver: m_messageReceiverServices){
-				messageReceivers.put(messageReceiver.getServiceName(), messageReceiver);
+				if (messageReceivers.containsKey(messageReceiver.getClientInstanceId())) 
+					throw new RuntimeException("duplicate messageReceiver '"+messageReceiver.getClientInstanceId()
+							+ "' in configuration");
+				messageReceivers.put(messageReceiver.getClientInstanceId(), messageReceiver);
 			}
 		}
 
@@ -118,8 +151,10 @@ public class Controller {
 			if (m_clientMap.containsKey(receiverConfig.getClientInstanceId())) 
 				throw new RuntimeException("duplicate ClientInstanceId '"+receiverConfig.getClientInstanceId()
 						+ "' in configuration");
+			
 			MqttRxService rxService = messageReceivers.get(receiverConfig.getClientInstanceId());
-			if(rxService!=null && rxService.getServiceType().equals(receiverConfig.getClientType())){
+			
+			if(rxService!=null && rxService.getClientType().equals(receiverConfig.getClientType())){
 				if(receiverConfig.getConfiguration()!=null) {
 					Map<String, String> properties = new HashMap<String, String>();
 					for(ConfigProperty prop:receiverConfig.getConfiguration()){
@@ -137,65 +172,88 @@ public class Controller {
 		}
 
 		// set up message queue
-		m_messageNotificationClientQueueImpl = new MessageNotificationClientQueueImpl();
 		m_messageNotificationClientQueueImpl.setMaxMessageQueueLength(m_MQTTReceiverConfig.getMaxMessageQueueLength());
+		m_messageNotificationClientQueueImpl.setMaxMessageQueueThreads(m_MQTTReceiverConfig.getMaxMessageQueueThreads());
+
+		m_messageNotificationClientQueueImpl.setNotificationHandlingClients(Arrays.asList());
 
 		// add m_messageReceiverServices to message queue
 		// used to add rest interface
 		List<MessageNotifier> messageNotifiers = new ArrayList<MessageNotifier>();
 		for(String clientInstanceId : m_rxServiceMap.keySet()){
-			LOG.debug("adding rxService client:"+clientInstanceId);
+			LOG.debug("adding rxService to message queue client:"+clientInstanceId);
 			messageNotifiers.add(m_rxServiceMap.get(clientInstanceId));
 		}
 
 		// add user defined mqtt receivers to message queue
 		for(String clientInstanceId : m_clientMap.keySet()){
-			LOG.debug("adding mqtt client:"+clientInstanceId);
+			LOG.debug("adding mqtt to message queue client:"+clientInstanceId);
 			messageNotifiers.add(m_clientMap.get(clientInstanceId));
 		}
 
 		m_messageNotificationClientQueueImpl.setMessageNotifiers(messageNotifiers);
 
-		Map<String, NotificationClient> topicHandlingClients = new LinkedHashMap<String, NotificationClient>();
-
-		// set up event collectors
-		for(MessageEventParserConfig messageEventParserConfig : m_MQTTReceiverConfig.getMessageEventParsers()){
-			// create json parser
-			XmlGroups xmlGroups = messageEventParserConfig.getXmlGroups();
-			OnmsAttributeMessageHandler onmsAttributeMessageHandler = new OnmsAttributeMessageHandler(xmlGroups);
-
-			MqttEventNotificationClient eventClient = new MqttEventNotificationClient();
-			//TODO SET UP CONFIG 
-			//eventClient.setOnmsAttributeJsonHandler(onmsAttributeJsonHandler);
-
-			for(String eventSubscriptionTopic:messageEventParserConfig.getSubscriptionTopics()){
-				LOG.debug("adding handler for eventSubscriptionTopic:"+eventSubscriptionTopic);
-				if(topicHandlingClients.containsKey(eventSubscriptionTopic)) {
-					LOG.error("duplicate handler for eventSubscriptionTopic:"+eventSubscriptionTopic);
-				} else 	topicHandlingClients.put(eventSubscriptionTopic, eventClient);
+		Map<String, MessageParserConfig> topicDataParserMap = new HashMap<String, MessageParserConfig>();
+		for( MessageDataParserConfig messageDataParser:m_MQTTReceiverConfig.getMessageDataParsers()){
+			for(String subscription:messageDataParser.getSubscriptionTopics()){
+				if (topicDataParserMap.containsKey(subscription)) 
+					throw new RuntimeException("duplicate data topic subscription '"+subscription
+							+ "' in configuration");
+				topicDataParserMap.put(subscription, messageDataParser);
 			}
 		}
-
-		// set up pm collectors
-		for( MessageDataParserConfig messageDataParserConfig : m_MQTTReceiverConfig.getMessageDataParsers()){
-			// create json parser
-			XmlGroups xmlGroups = messageDataParserConfig.getXmlGroups();
-			OnmsAttributeMessageHandler onmsAttributeMessageHandler = new OnmsAttributeMessageHandler(xmlGroups);
-
-			MqttDataNotificationClient dataClient = new MqttDataNotificationClient();
-			//TODO SET UP CONFIG 
-			//dataClient.setOnmsAttributeJsonHandler(onmsAttributeJsonHandler);
-
-			for(String dataSubscriptionTopic:messageDataParserConfig.getSubscriptionTopics()){
-				LOG.debug("adding handler for performanceSubscriptionTopic:"+dataSubscriptionTopic);
-				topicHandlingClients.put(dataSubscriptionTopic, dataClient);
-			}
-		}
-
-		m_messageNotificationClientQueueImpl.setTopicHandlingClients(topicHandlingClients);
 		
-		// init receivers
+		Map<String, MessageParserConfig> topicEventParserMap = new HashMap<String, MessageParserConfig>();
+		for( MessageDataParserConfig messageDataParser:m_MQTTReceiverConfig.getMessageDataParsers()){
+			for(String subscription:messageDataParser.getSubscriptionTopics()){
+				if (topicEventParserMap.containsKey(subscription)) 
+					throw new RuntimeException("duplicate event topic subscription '"+subscription
+							+ "' in configuration");
+				topicEventParserMap.put(subscription, messageDataParser);
+			}
+		}
+		
+		//TODO allow events and data from same topic
+		
+		// check no duplicate topics
+		for(String subscription: topicEventParserMap.keySet()){
+			if (topicDataParserMap.containsKey(subscription))throw new RuntimeException("duplicate event topic subscription '"+subscription
+					+ "' for events and data in configuration");
+			topicDataParserMap.put(subscription, topicEventParserMap.get(subscription));
+		}
+		
+		m_notificationMessageHandler.setTopicParserMap(topicDataParserMap);
 
+	}
+
+	public void init(){
+		// load configuration and set up all classes
+		loadConfig();
+
+		// initialise nodeCache
+		m_nodeByForeignSourceCacheImpl.init();
+
+		// initialise messageQueue
+		m_messageNotificationClientQueueImpl.init();
+
+		// initialise receivers
+		for(String clientId: m_clientMap.keySet()){
+			LOG.info("initialising"+clientId);
+			try{
+			m_clientMap.get(clientId).init();
+			}catch (Exception ex){
+				LOG.error("problem initialising clientId:"+clientId,ex);
+			}
+		}
+		
+		for(String clientId: m_rxServiceMap.keySet()){
+			LOG.info("initialising"+clientId);
+			try{
+				m_rxServiceMap.get(clientId).init();
+			}catch (Exception ex){
+				LOG.error("problem initialising clientId:"+clientId,ex);
+			}
+		}
 
 	}
 
@@ -205,25 +263,37 @@ public class Controller {
 		for (String clientInstanceId : m_clientMap.keySet()){
 			try {
 				MQTTClientImpl client = m_clientMap.get(clientInstanceId);
-				client.destroy();
+				if(client!=null) client.destroy();
 				m_clientMap.remove(clientInstanceId);
 			} catch (Exception e){
 				LOG.error("problem destroying client "+clientInstanceId, e);
 			}
 		}
-		
+
 		// disconnect and destroy MqttRxService receivers
 		for (String clientInstanceId : m_rxServiceMap.keySet()){
 			try {
 				MqttRxService client = m_rxServiceMap.get(clientInstanceId);
-				client.destroy();
+				if(client!=null) client.destroy();
 				m_rxServiceMap.remove(clientInstanceId);
 			} catch (Exception e){
 				LOG.error("problem destroying m_rxServiceMap client "+clientInstanceId, e);
 			}
 		}
 
-		//m_nodeByForeignSourceCacheImpl.destroy(); // TODO
+		// destroy message queue (shuts down consumer threads)
+		try {
+			m_messageNotificationClientQueueImpl.destroy();
+		} catch (Exception e){
+			LOG.error("problem destroying m_messageNotificationClientQueueImpl", e);
+		}
+
+		// destroy node cache
+		try{
+			m_nodeByForeignSourceCacheImpl.destroy();
+		} catch (Exception e){
+			LOG.error("problem destroying m_nodeByForeignSourceCacheImpl", e);
+		}
 
 	}
 
@@ -257,4 +327,6 @@ public class Controller {
 
 
 	}
+
+
 }
